Move InterpolatedProperty handling outside of StyleSetEvaluator.
